Add in-memory storage option to process_plink()

+++ R/process_plink.R
@@ -3,17 +3,17 @@
 #' @param data_dir              The path to the bed/bim/fam data files, *without* a trailing "/" (e.g., use `data_dir = '~/my_dir'`, **not** `data_dir = '~/my_dir/'`)
 #' @param data_prefix           The prefix (as a character string) of the bed/fam data files (e.g., `data_prefix = 'mydata'`)
 #' @param rds_dir               The path to the directory in which you want to create the new '.rds' and '.bk' files. Defaults to `data_dir`
-#' @param rds_prefix            String specifying the user's preferred filename for the to-be-created .rds file (will be create insie `rds_dir` folder)
+#' @param rds_prefix            String specifying the user's preferred filename for the to-be-created .rds file (will be create inside `rds_dir` folder). If no rds_prefix is provided, the processed data files will be returned in memory.
 #'                              Note: 'rds_prefix' cannot be the same as 'data_prefix'
 #' @param logfile               Optional: the name (character string) of the prefix of the logfile to be written in 'rds_dir'. Default to NULL (no log file written).
 #'                              Note: if you supply a file path in this argument, it will error out with a "file not found" error. Only supply the string; e.g., if you want my_log.log, supply 'my_log', the my_log.log file will appear in rds_dir.
 #' @param impute                Logical: should data be imputed? Default to TRUE.
 #' @param impute_method         If 'impute' = TRUE, this argument will specify the kind of imputation desired. Options are:
-#'                                * mode (default): Imputes the most frequent call. See `bigsnpr::snp_fastImputeSimple()` for details.
-#'                                * random: Imputes sampling according to allele frequencies.
-#'                                * mean0: Imputes the rounded mean.
-#'                                * mean2: Imputes the mean rounded to 2 decimal places.
-#'                                * xgboost: Imputes using an algorithm based on local XGBoost models. See `bigsnpr::snp_fastImpute()` for details. Note: this can take several minutes, even for a relatively small data set.
+#' * mode (default): Imputes the most frequent call. See `bigsnpr::snp_fastImputeSimple()` for details.
+#' * random: Imputes sampling according to allele frequencies.
+#' * mean0: Imputes the rounded mean.
+#' * mean2: Imputes the mean rounded to 2 decimal places.
+#' * xgboost: Imputes using an algorithm based on local XGBoost models. See `bigsnpr::snp_fastImpute()` for details. Note: this can take several minutes, even for a relatively small data set.
 
 #' @param id_var              String specifying which column of the PLINK `.fam` file has the unique sample identifiers. Options are "IID" (default) and "FID"
 #' @param parallel            Logical: should the computations within this function be run in parallel? Defaults to TRUE. See `count_cores()` and `?bigparallelr::assert_cores` for more details.
@@ -49,7 +49,7 @@
 process_plink <- function(data_dir,
                           data_prefix,
                           rds_dir = data_dir,
-                          rds_prefix,
+                          rds_prefix = NULL,
                           logfile = NULL,
                           impute = TRUE,
                           impute_method = 'mode',
@@ -140,43 +140,56 @@ process_plink <- function(data_dir,
     col = 1:ncol(step3$genotypes),
     type = 'double', # this is why we're making a copy -- need type 'double' downstream
     backingpath = rds_dir,
-    backingfile = paste0(rds_prefix, '.bk'),
-    descriptorfile = paste0(rds_prefix, '.desc')
+    backingfile = paste0('processed_', data_prefix, '.bk'),
+    descriptorfile = paste0('processed_', data_prefix, '.desc')
   )
 
   ret <- structure(list(X = bigmemory::describe(X),
-              map = step3$map,
-              fam = step3$fam,
-              n = step3$n,
-              p = step3$p), class = "processed_plink")
+                        map = step3$map,
+                        fam = step3$fam,
+                        n = step3$n,
+                        p = step3$p), class = "processed_plink")
 
-  rds_filename <- paste0(rds_prefix, ".rds")
-  saveRDS(ret, file = file.path(rds_dir, rds_filename))
+  # handle return format -------------------------------------------------------
+
+  if(!is.null(rds_prefix)) {
+    rds_filename <- paste0(rds_prefix, ".rds")
+    saveRDS(ret, file = file.path(rds_dir, rds_filename))
+  } else {
+    # TODO: add warning if object size is greater than some threshold
+    ret$X <- bigmemory::attach.big.matrix(ret$X)[,]
+
+    list.files(rds_dir, pattern = paste0('processed_', data_prefix, '.bk'),
+               full.names = TRUE) |>
+      file.remove()
+    list.files(rds_dir, pattern = paste0('processed_', data_prefix, '.desc'),
+               full.names = TRUE) |>
+      file.remove()
+  }
 
   # cleanup --------------------------------------------------------------------
+
   # These steps remove intermediate rds/bk files created by the steps of the data management process
   list.files(rds_dir, pattern=paste0('^file.*.bk'), full.names=TRUE) |>
     file.remove()
   list.files(rds_dir, pattern = paste0('^', data_prefix, ".rds"), full.names = TRUE) |>
     file.remove()
   list.files(rds_dir, pattern = paste0('^', data_prefix, ".bk"), full.names = TRUE) |>
-    file.remove()
+    file.remove() #TODO: Determine the permissions error origin
 
   gc()
 
+  if(!quiet){cat("\nprocess_plink() completed")}
 
-  if(!quiet){cat("\nprocess_plink() completed \nProcessed files now saved as",
-                 file.path(rds_dir, rds_filename))}
+  if(!is.null(rds_prefix)) {
+    cat("\nProcessed files now saved as",
+        file.path(rds_dir, rds_filename))
 
-  cat("\nprocess_plink() completed. \nProcessed files now saved as",
-      file.path(rds_dir, rds_filename),
-      "at", pretty_time(),
-      file = logfile, append = TRUE)
-
-  # file.remove(tempfile()) # TODO: Not needed? Think about best way to handle logfile...
-  return(file.path(rds_dir, rds_filename))
+    # file.remove(tempfile())
+    # TODO: Not needed? Think about best way to handle logfile...
+    return(file.path(rds_dir, rds_filename))
+  } else {
+    return(ret)
+  }
 
 }
-
-
-
